minify script injection at build-time
