+++ package.json
@@ -28,7 +28,7 @@
     "dist"
   ],
   "scripts": {
-    "build": "pnpm dev:prepare && nuxt-module-build",
+    "build": "pnpm dev:prepare && nuxt-module-build && esbuild --minify src/script.ts --outfile=dist/script.min.js",
     "dev": "nuxi dev playground",
     "dev:build": "nuxi build playground",
     "dev:prepare": "pnpm nuxt-module-build --stub && nuxi prepare playground",
@@ -60,6 +60,7 @@
     "@vitest/coverage-v8": "0.34.6",
     "@vue/test-utils": "2.4.4",
     "bumpp": "9.2.1",
+    "esbuild": "^0.20.1",
     "eslint": "8.57.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-prettier": "5.1.3",
@@ -70,6 +71,7 @@
     "lint-staged": "15.2.2",
     "magic-regexp": "0.7.0",
     "nuxt": "3.10.0",
+    "nuxt-time": "workspace:^",
     "playwright": "1.40.1",
     "prettier": "3.2.5",
     "typescript": "5.1.6",
