+++ src/runtime/components/NuxtTime.vue
@@ -1,5 +1,7 @@
 <script setup lang='ts'>
 import { computed, getCurrentInstance, useNuxtApp, useHead } from '#imports'
+import scriptString from '../../script'
+
 
 const props = withDefaults(defineProps<{
   locale?: string
@@ -58,32 +60,13 @@ if (process.server) {
       dataset[`data-${propInKebabCase}`] = props?.[prop as keyof typeof props]
     }
   }
-  useHead({
-    script: [{
-      tagPosition: 'bodyClose',
-      key: 'nuxt-time',
-      innerHTML: `
-        document.querySelectorAll('[data-n-time]').forEach(el => {
-          const toCamelCase = (name, index) => {
-            if (index > 0) { return name[0].toUpperCase() + name.slice(1) };
-            return name;
-          };
-
-          const date = new Date(el.getAttribute('datetime'));
-          const options = {};
-          for (const name of el.getAttributeNames()) {
-            if (name.startsWith('data-')) {
-              const optionName = name.slice(5).split('-').map(toCamelCase).join('');
-              options[optionName] = el.getAttribute(name);
-            }
-          }
-
-          const formatter = new Intl.DateTimeFormat(options.locale, options);
-          el.textContent = formatter.format(date)
-        })
-      `.replace(/\s+/g, ' ')
-    }]
-  })
+ useHead({
+  script: [{
+    tagPosition: 'bodyClose',
+    key: 'nuxt-time',
+    innerHTML: scriptString,
+  }]
+});
 }
 </script>
